Split-up pds view tests
